seprate lazy vim package to files.
