+++ init.lua
@@ -8,7 +8,7 @@ vim.filetype.add {
   },
 }
 
-require 'plugins'
+require 'core.lazy'
 
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
