+++ lua/plugins/autopairs.lua
@@ -1,6 +1,3 @@
--- autopairs
--- https://github.com/windwp/nvim-autopairs
-
 return {
   'windwp/nvim-autopairs',
   event = 'InsertEnter',
