+++ lua/plugins/indent_blankline.lua
@@ -1,0 +1,5 @@
+return {
+  'lukas-reineke/indent-blankline.nvim',
+  main = 'ibl',
+  opts = {},
+}
