+++ lua/plugins/lazygit.lua
@@ -1,0 +1,16 @@
+return {
+  'kdheepak/lazygit.nvim',
+  cmd = {
+    'LazyGit',
+    'LazyGitConfig',
+    'LazyGitCurrentFile',
+    'LazyGitFilter',
+    'LazyGitFilterCurrentFile',
+  },
+  dependencies = {
+    'nvim-lua/plenary.nvim',
+  },
+  keys = {
+    { '<leader>gg', '<cmd>LazyGit<cr>', desc = 'LazyGit' },
+  },
+}
