+++ lua/plugins/ruby.lua
@@ -1,0 +1,4 @@
+return {
+  { 'tpope/vim-bundler', ft = { 'ruby', 'eruby' } },
+  { 'tpope/vim-endwise', ft = { 'ruby', 'eruby' } },
+}
