+++ lua/plugins/tokyonight.lua
@@ -1,0 +1,8 @@
+return {
+  'folke/tokyonight.nvim',
+  priority = 1000, -- Make sure to load this before all the other start plugins.
+  init = function()
+    vim.cmd.colorscheme 'tokyonight-storm'
+    vim.cmd.hi 'Comment gui=none'
+  end,
+}
