+++ lua/plugins/which-key.lua
@@ -1,0 +1,18 @@
+return {
+  'folke/which-key.nvim',
+  event = 'VimEnter',
+  opts = {
+    icons = {
+      mappings = vim.g.have_nerd_font,
+    },
+    spec = {
+      { '<leader>c', group = '[C]ode', icon = '', mode = { 'n', 'x' } },
+      { '<leader>d', group = '[D]ocument', icon = '󰈙' },
+      { '<leader>g', group = '[G]it', icon = '󰊢' },
+      { '<leader>r', group = '[R]ename', icon = '󰑕' },
+      { '<leader>s', group = '[S]earch', icon = '' },
+      { '<leader>w', group = '[W]orkspace', icon = '󰯉' },
+      { '<leader>t', group = '[T]oggle', icon = '' },
+    },
+  },
+}
